+++ dwds/lib/src/injected/client.js
@@ -26213,25 +26213,30 @@
               t4 = W.Event;
               W._EventStreamSubscription$(client._eventSource, "error", H.functionTypeCheck(t2.get$addError(), {func: 1, ret: -1, args: [t4]}), false, t4);
               client._startPostingMessages$0();
-              $async$goto = J.$eq$(self.$dartModuleStrategy, "require") ? 2 : 4;
-              break;
+              t4 = new W._EventStream(client._eventSource, "open", false, [t4]);
+              $async$goto = 2;
+              return P._asyncAwait(t4.get$first(t4), $async$call$0);
             case 2:
+              // returning from await.
+              $async$goto = J.$eq$(self.$dartModuleStrategy, "require") ? 3 : 5;
+              break;
+            case 3:
               // then
-              $async$goto = 5;
+              $async$goto = 6;
               return P._asyncAwait(X.RequireRestarter_create(), $async$call$0);
-            case 5:
+            case 6:
               // returning from await.
               restarter = $async$result;
               // goto join
-              $async$goto = 3;
+              $async$goto = 4;
               break;
-            case 4:
+            case 5:
               // else
               if (J.$eq$(self.$dartModuleStrategy, "dart_library.import"))
                 restarter = new Z.LegacyRestarter();
               else
                 throw H.wrapException(P.StateError$("Unknown module strategy: " + H.S(self.$dartModuleStrategy)));
-            case 3:
+            case 4:
               // join
               manager = new Q.ReloadingManager(client, self.$dartAppId, self.$dartAppInstanceId, restarter);
               t4 = P.allowInterop(new D.main__closure(manager), {func: 1, ret: [S.Promise, -2]});